+++ src/components/Select/__demo__/popoverProps.jsx
@@ -11,7 +11,7 @@ class Demo extends React.Component {
             <div>
                 {animations.map(animation => (
                     <div className="demo-wrap" key={animation}>
-                        <Select defaultValue={1} popover={{ animation }}>
+                        <Select defaultValue={1} popoverProps={{ animation }}>
                             <Option value={1}>1</Option>
                             <Option value={2}>2</Option>
                             <Option value={3}>3</Option>