+++ src/components/ThemeProvider/ocean.js
@@ -8,5 +8,13 @@ export default {
         primary5: '#c3caf7',
         primary6: '#fcfcfd',
         primary7: '#cae3ff'
+    },
+    materialVars: {
+        primaryBoxShadow:
+            '0 1px 5px 0 rgba(21,56,195,.12), 0 2px 2px 0 rgba(8, 39, 156,.24), 0 3px 1px -3px rgba(58, 96, 245,.34)',
+        primaryBoxShadowActive:
+            '0 1px 10px 0 rgba(21,56,195,.12), 0 2px 4px -1px rgba(21,56,195,.2), 0 4px 5px 0 rgba(21,56,195,.14)',
+        primaryLinearGradient: 'linear-gradient(#415bf5,#415bf5)',
+        primaryLinearGradientActive: 'linear-gradient(#3d5aff,#3d5aff)'
     }
 };